requirements.txt
pandas
scikit-learn
nltk
fastapi
uvicorn
pydantic
transformers
torch
google-generativeai
python-dotenv
accelerate
